refactor: split high-conflict monoliths into per-domain files
Three structural splits to reduce conflict surface in the open PR queue.
Pure mechanical extraction — no behavior change.

## ConditionPatterns.kt → +BaselineDeviationPatterns.kt
Moves the 11 inline founding patterns (infection_onset, sleep_disruption,
cardiovascular_stress, overtraining, metabolic_drift,
cardiorespiratory_deconditioning, chronic_inflammation, recovery_deficit,
respiratory_infection, mental_health_correlate, menstrual_cycle_anomaly)
out into a sibling file matching the existing convention
(SepsisScreenPattern, AfibRhythmPattern, PregnancyPatterns, …).

ConditionPatterns.kt is now ~140 lines (was 497), holding only the
ConditionPattern + SignalRule data classes + enums + the `all`
aggregator. New baseline-deviation patterns land in
BaselineDeviationPatterns.kt; new specialty groups get their own file.

Test references updated from ConditionPatterns.X to
BaselineDeviationPatterns.X (4 test files).

## SettingsScreen.kt → +IdentityCard.kt
Extracts the "Identity" card (12 owner-context rows) into its own
composable. New self-surfaces add one nav-callback param + one row in
IdentityCard.kt, which no other PR touches.

SettingsScreen.kt now 490 lines (was 500) — 10 lines of headroom on the
500-line cap.

## MainActivity.kt → +IdentityRoutes.kt
Extracts 11 simple back-only routes (risk_profile, physiology_state,
frail_assessment, goals_of_care, headache_diary, fast_stroke,
esas_capture, traditional_medicine_context, medications,
immunisations, preventive_care) as a NavGraphBuilder extension.

MainActivity.kt now 444 lines (was 474). Stateful routes (event-sheet
coupled, deep-link argument routes) stay in MainActivity since
extraction would require plumbing local state setters.

## Why now
The 11-PR audit-derived merge queue (#214-#233) has been hitting these
three files repeatedly. Every alerts PR conflicts in ConditionPatterns;
every "Self" PR conflicts in MainActivity + SettingsScreen. Landing
this refactor first means downstream PRs add to dedicated landing
zones instead of fighting the 500-line cap in the shared monoliths.
